refactor: cardNumbers 로직 분리
